parser wip

+++ source/types.h
@@ -3,8 +3,8 @@
 #include <cstdint>
 
 using byte = uint8_t;
-using integer = int64_t;
-using number = double;
+using integer = int32_t;
+using number = float;
 
 using utf16unit = char16_t;
 using character = char32_t;
